+++ client/src/App.css
@@ -1,42 +1,3 @@
-/* *{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-.Nav{
-  height: 8vh;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  background: linear-gradient(150deg,purple,pink);
-  position: sticky;
-  top:0;
-  left:0;
-}
-
-.Nav a{
-  color: white;
-  text-decoration: none;
-}
-
-.rightNav ul{
-  width:40vw;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  list-style: none;
-}
-
-.rightNav li{
-  padding: 10px;
-  transition: all 0.5s ease-in-out;
-}
-
-.rightNav li:hover{
-  background: rgb(88, 39, 4);
-} */
-
-/* ////////////////////////////// */
 
 body {
   margin: 0;
@@ -50,37 +11,6 @@ body {
   padding: 0;
   box-sizing: border-box;
 }
-
-/*
-.Nav{
-  height: 8vh;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  background: linear-gradient(150deg,blue, skyblue);
-  position: sticky;
-  top:0;
-  left:0;
-}
-.Nav a{
-  color: white;
-  text-decoration: none;
-}
-.rightNav ul{
-  width:40vw;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  list-style: none;
-}
-.rightNav li{
-  padding: 10px;
-  transition: all 0.5s ease-in-out;
-}
-.rightNav li:hover{
-  background: rgb(88, 39, 4);
-}
-*/
 
 /* CORE STYLES */
 :root {
@@ -219,149 +149,6 @@ body {
   transform: rotate(225deg);
 }
 
-<<<<<<< HEAD
-/* *{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-.Nav{
-  height: 8vh;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  background: linear-gradient(150deg,purple,pink);
-  position: sticky;
-  top:0;
-  left:0;
-}
-
-.Nav a{
-  color: white;
-  text-decoration: none;
-}
-
-.rightNav ul{
-  width:40vw;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  list-style: none;
-}
-
-.rightNav li{
-  padding: 10px;
-  transition: all 0.5s ease-in-out;
-}
-
-.rightNav li:hover{
-  background: rgb(88, 39, 4);
-} */
-
-/* ////////////////////////////// */
-
-body {
-  margin: 0;
-  padding: 0;
-  background: url("food.jpg") no-repeat center center fixed;
-  font-family: sans-serif;
-}
-
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-/*
-.Nav{
-  height: 8vh;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  background: linear-gradient(150deg,blue, skyblue);
-  position: sticky;
-  top:0;
-  left:0;
-}
-.Nav a{
-  color: white;
-  text-decoration: none;
-}
-.rightNav ul{
-  width:40vw;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  list-style: none;
-}
-.rightNav li{
-  padding: 10px;
-  transition: all 0.5s ease-in-out;
-}
-.rightNav li:hover{
-  background: rgb(88, 39, 4);
-}
-*/
-
-/* CORE STYLES */
-:root {
-  --primary-color: rgba(13, 110, 139, 0.75);
-  --overlay-color: rgba(24, 39, 51 , 0.85);
-  --menu-speed: 0.75s;
-}
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: 'Roboto', sans-serif;
-  line-height: 1.4;
-}
-
-.container {
-  max-width: 960px;
-  margin: auto;
-  overflow: hidden;
-  padding: 0 3rem;
-}
-
-.showcase .showcase-inner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  height: 100%;
-}
-
-.showcase h1 {
-  font-size: 4rem;
-}
-
-.showcase p {
-  font-size: 1.3rem;
-}
-
-.btn {
-  display: inline-block;
-  border: none;
-  background: var(--primary-color);
-  color: #fff;
-  padding: 0.75rem 1.5rem;
-  margin-top: 1rem;
-  transition: opacity 1s ease-in-out;
-  text-decoration: none;
-}
-
-.btn:hover {
-  opacity: 0.7;
-}
-
-=======
 /* Show Menu */
 .menu-wrap .toggler:checked ~ .menu {
   visibility: visible;
@@ -423,4 +210,3 @@ body {
   text-decoration: none;
   transition: color 0.4s ease;
 }
->>>>>>> aeb0a94c9da26ac28ad59ba0490b547951067ddc
